refactor: melhorando organizacao do codigo e modificando alguns detalhes de ui
